fix: zero-trust audit 12-14 remediation - dedupe inferred/unknown/link/bonus/amend/trace + date-delegation + forecast breached-baseline
- timeline: dedupe inferred on (source,day) 116->0, unknown-reserve via median, scheduled bonus guard, linked lifecycle dedupe, amend-aware history, per-flow FX trace fields
- message_income: delegate forecast_end/clamp to temporal (remove duplicate calendar logic)
- forecast: early-return breached baseline + floored hi + post-assert (concurrent fix kept)
- state/eligibility/decision/optimizer/spending_changes: doc clarifications, Decimal annotation
- output.csv: 19 amount_safe uplifts conservative->accurate, status mix 33/182/29/6 preserved, validator PASS
- tests: add sections 15-21 gap coverage (21 tests), total 197 passed
- code.zip: rebuilt 93 entries

+++ src/affordai/decision/decision.py
@@ -2,6 +2,7 @@
 from __future__ import annotations
 
 from dataclasses import dataclass, field
+from decimal import Decimal
 
 
 @dataclass
@@ -9,7 +10,7 @@ class Decision:
     original_index: int
     request_id: str
     user_id: str
-    amount_safe_to_pay: float
+    amount_safe_to_pay: Decimal
     affordability_status: str
     recommended_payment_method: str
     payment_plan: str
